% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrSurvGSD-package.R
\docType{package}
\name{CorrSurvGSD-package}
\alias{CorrSurvGSD}
\alias{CorrSurvGSD-package}
\title{CorrSurvGSD: Closed-Form Evaluation of Correlated Group Sequential Designs}
\description{
Closed-form calculations for correlated PFS and OS group sequential designs under the Fleischer model, with theoretical power calculations and simulation utilities.
}
\author{
\strong{Maintainer}: Xingjian Ma \email{xingjian.ma@uconn.edu}

}
\keyword{internal}
