% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the CorrSurvGSD Shiny application.}
\usage{
run_app(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\value{
The result returned by \code{\link[shiny:runApp]{shiny::runApp()}}.
}
\description{
Run the CorrSurvGSD Shiny application.
}
